Centering everything, the loading.gif also
centering new email account form horizontally and vertical adjustments also

This is a combination of 6 commits.
centering the <new email> form

stupid error in blind css

changing the structure of the markdown arround  <new email> form for it to be included into the #app-content div, and therefor it can be centered while keeping absolute positionning, in respect with the rest of the cascading css structure of the app

CSS adaptation of the new email account form

adaptation of the css structure now that content is truly relatively positionned; removes a bit of top margin to now have scroll in the form on small heighed screens

switching to a cleaner aligning method

now adding padding to the top to 'center' the form vertically

also centering the loading.gif to the form, forgot to do so.

syntax correction based on Codacy