Make redirect in Firefox work again and harden CSP policy
This makes the redirect in Firefox work again by using CSP instead of a sandboxed iframe. Unit tests are coming later because the message controller has some hard-coded references to other classes with side-effect such as `\OCA\Mail\Account`. Trying to come up with a fix for that as well in another PR.

Fixes https://github.com/owncloud/mail/issues/206