log SMTP communication #1392
* log SMTP communication #1392

I copied the thing form imap and changed it. After I try it I noticed it would be better to split the logs. So one for imap and one for smtp.

* additional to my commit form log SMTP communication

Necessary changes for the issue_template that would incur by splitting the horde logs.

* Update issue_template.md

* additional to my commit form log SMTP communication

Necessary changes for the readme that would incur by splitting the horde logs.