automatically sign and package the app
This rsyncs the files into a directory that can then be used to sign the app's
files. Moreover, the created tar.gz is then signed and the certificate is
printed to the command line. All steps that require the private key are only
run if the key is available. Therefore peole can also use this to build
their own version :-)