Fix the account selected as reply account
The code to determine the right account was already in place,
but it used a global jQuery selector. Since the view has not
been attached to the DOM at that point, the selector doesn't
select anything and the line of code has no effect.

Fixes https://github.com/nextcloud/mail/issues/10

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>