Add IMAP integration test framework
Add a trait-based testing framework that automatically resets
the IMAP testing account before a test case is run to prevent
side effects and dependencies across different tests.
A simple self test was added to ensure the functionality works
as expected.
A new super class for integration tests was added which should
be used now instead of phpunit's class. This will allow to write
tests with less boilerplate code, as we can for example check
the usage of the IMAP test trait and reset accounts automatically.
This was inspired by the Laravel testing capabilities.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>