Fix mailto routing
If the composer is shown for a mailto handler then we have to load the first
account and first folder before we actually render the composer view. Since
that is async and we don't know when it will complete, the composer vanishes
again when the first message of the folder is shown. Hence a promise-based
approach was chosen here to propagate the loading of the folder messages back
to the routing controller to be able to wait for the folder to have loaded and
then trigger to show the composer.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>