Fix selection of next message in unified inbox
If the user deletes a message in the unified inbox, the next (previous)
message is selected automatically. In case that message actually is part
of a different one than the deleted one an error was shown because the
new message could not be loaded. This happened because the wrong account
and folder objects were passed to the 'load' event.

Fixes https://github.com/nextcloud/mail/issues/465

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>