Add earlier loading feedback
In terms of UX, I like to know what the app is doing, thus
IMO it makes sense to tell the user we're loading their accounts.

From a technical perspective, I know that the app's js has started
executing. This is usefull for debugging where you didn't know
whether the page was still loading or the app was already started.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>