Fix active message selection
The state that determines whether a messages is actively shown
or not is stored as a property of the message object. Since
we now use the very same message data in the unified inbox
as in specific ones, a message marked as being active kept its
state when the users switched to the unified inbox. Now all states
are reset to ensure we only have one active message.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>