Fix encoding issues of the folder display name
Folders (technically speaking: mailboxes) have utf7/utf8 names. When we
encode the ID for the client, we use base64. The decoding will fail as
JavaScript assumes that base64 encoded strings are actually ASCII.
Instead of rewriting all the ID handling to not use base64, I decided to
just add a new prop with the unencoded displayname. Then it stays utf8
all the time.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>