Add individually sync'ed new messages to the unified mailbox
In the background we invoke the synchronization of all individual
inboxes. With a recent change in the Vuex logic we didn't push the new
message to the unified inbox as well. So you could see a notitification
about new messages but when you switched to the Mail tab it wasn't
there. You first had to switch to the individual mailbox and then back
to get the latest entries.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>