Do not trigger scroll event on page load
vue-infinite-scroll can detect when the element is too big to fit the
contents on first render. This is typically the case for us as the
message list is bigger than the container. However, we don't want to
trigger the loading of the next page yet, as we first have to fetch the
initial one.
Luckily there is an option to disable the immediate check.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>