Fix the UID validity change logic
UID validity is a way for the IMAP server to signal that the UIDs - that
are usually persistend - have changed. Whenever this happens clients
have to wipe their cached data and read the IMAP data anew. Since making
the initial sync process resumable, it does not clear the messages of a
mailbox as these might be from a previous sync attempt. Hence, we have
to manually clear the cache on a UID validity change before we invoke
the initial sync routine.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>