Fix endless initial sync due to empty partial page
The initial message cache sync estimates the UID range of the next
message page so it can fetch messages efficiently. There was an
uncovered edge case where the next page might not return any message,
simple because there is a larger gap between the existing UIDs.
Therefore the sync process was aborted as incomplete over and over as it
could not make any progress and the number of known messages never
reached the total number of messages on IMAP. This patch changes the
logic so it retries the next page if the current one is empty. To break
a possibly infinite recursion there is a check about reaching the
highest UID reported by the server.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>