Cache no avatar available in memcache as well
We cache the HTTP response for when no avatar is available, so the same
browser won't ask for this for the next few hours. However, this doesn't
work across devices, so a tiny bit of work is spent on a task that can
be cached.

This is most noticeable in dev setup where the HTTP cache is disabled.
The server had to always check if avatars are available. This is quite
slow and causes lots of requests to queue.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>