Fix unseen messages loading infinitely
When an unseen message is opened, we flag it as seen. So far, so good.
But in this code path we returned early from the `fetchMessage`
function, meaning that `this.loading` remained set to `true` and the
loading spinner never vanished.

My guess is that this is a leftover from the migration from promises to
async-await.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>