Fix signature detection and signature design
We can't assume line endings of specific platforms but should be
flexible with that when detecting the signature part of an email.

Also this fixes the styling of signatures a bit. They looked quite off
now that the distinct styling came back.

Fine-tuning in terms of design can be done in follow-ups. This PR is
mainly about getting the detection back.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>