Collapse plain text signatures into a detail element
Signatures are a bit of clutter when you view a conversation. So let's
collapse them and only show on click.

Luckily there is an HTML element we can use for this: `<details>`.

If the design should be improved then let's do a follow-up. This PR
focuses on the splittig logic.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>