Prevent account settings reuse when account changes
Vue is clever and always tries to recycle a component when certain
criteria is met. Apparently it also does it for components with objects
as props, which led to the problem that when you navigated from one
account's settings to another account's setting, the mailbox default
didn't update and you actually saw mailboxes of the previous account.

By key-ing the components we can express that the rendering depends on
the account ID and every time this value changes, Vue has to throw away
the instance and render a new one.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>