Fix initializing the editor mode
In the `data` of the Composer component we initialize the editor mode
with 'html' and update the mode as soon as the alias has been set (this
is done async). The uninitialized value is `-1` and not `undefined`,
hence before the patch the editor more always switch to `html` even when
the account was set to `plaintext`.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>