Check for method existence before calling during upgrade
The ``\OCA\Mail\Db\MessageMapper::findWithEmptyMessageId`` is new and
due to the nature of how Nextcloud upgrades work, the loaded code won't
have the new method when the post-migration repair step is run. Hence we
have to do a (arguably ugly) method exists check before the method is
called.

Fixes #4746

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>