Lower the chunk size for IMAP syncs
In #3110 I lowered the chances of too long IMAP commands by splitting
the sync process.

As we've seen in production it's still possible that certain accounts
run into the issue that their chunked UID list is producing an IMAP
command that is too long. So I'm once again lower the chances by
lowering the chunk size.

This could have a tiny effect on sync performance as it now takes more
chunks to check for the changed/vanished messages of one account.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>