Show inline attachments for plain text emails.
We distinguish between attachments (content-disposition: attachment) and inlineAttachments (content-disposition: inline).

For html messages we have a filter to replace a reference inside the message with an url to embed the attachment into the message.

For plain text messages it's not possible to embed / inline an attachment. In this case we merge inline attachments into attachments.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>