Calculate height for iframe by taggedElement
We disabled CSSTidy on purpose because some CSS rules were rewritten in a weird way: https://github.com/nextcloud/mail/pull/5424

For some reason the iframe height calculation does not work anymore. document.body.offsetHeight is 150.

taggedElement as heightCalculationMethod finds the bottom of the lowest element with a data-iframe-height attribute.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>