Make the sync response generic so it can hold two types of messages
We use the same container class for two purposes, once for the
IMAPMessage class, then for Message. Psalm rightfully complains that we
can't make assumptions of the returned polymorphic types, so it makes
sense to just type it as a generic class and let Psalm derive the
specialized type based on the constructor usage.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>