Fix initial sync stopping prematurely
The initial sync is broken up into chunks. These chunks are estimated
because we only know the lowest and highest UID and there are likely
gaps in this range. One of the last chunks might query a range that has
more messages than the allowed chunk size. We slice it down. This means
the actually feched max UID is lower than the theorical max UID for this
chunk.

If the stop condition is to stop at the theoretical max UID then we
occasionally stop too early. Where we actually have to stop is the
actually fetched highest UID.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>