Handle IMAP disconnects more explicitly
This changes the IMAP client usages to be used as a *resource* that is
freed after finished use. Previously we just memoized connections to the
same account to lower the number of connections, but that has still
shown to cause too many open connections during tests but possibly also
in production.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>