Make sure the account provisioning ID is an integer
If the database driver doesn't automagically cast the field to an int,
Nextcloud's entity mapper will take care of this.

Otherwise the field might be represented as string. This leads to an
wrong change detection when provisioned accounts are updated. The
assumption was that if the entity has no changes, we save the update.
Due to the type mispatch, there was still one updated field.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>