Fix memory leak when accessing Horde headers
Horde offers access to headers either unparsed, as stream or parsed into
an object. Before the patch we used the latter. That has the current bug
that it opens a PHP memory stream, parses into an object but leaves the
stream open. If this is done thousands of times, there is a considerate
amount of memory that doesn't get freed.

With the patch we request just the stream, trigger the parsing ourselves
and close the stream afterwards. This frees the memory.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>