Turn off cache for initial sync to prevent a memory leak
The cache implementation doesn't persist its values until the shutdown
of the process. For each iteration of the initial synchronization we
have to create one instance of the cache. Those two facts combined mean
that for each iteration Horde builds a new set of cache data for the
mailbox and therefore creates a memory leak.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>