Always serialize outbox message recipients into arrays
array_filter keeps the index of array elements prior to the filtering.
So if you have a message with one *to* and one *cc* values, the *to*
will have index 0 and *cc* has 1. json_encode will later consider the
former as a simple PHP array and map it to a json array. Since the *cc*
value has a non-0 index for its first element, it would convert the
value to a json object and use the 1 index as key.

By running the values through array_value we can re-index all recipient
arrays before they run through json_encode.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>