Prevent causal read for outbox message and recipient inserts
The message and the recipients are inserted in one transaction but the
recipients are read another time outside a transaction. Read-write split
database clusters might not be in full sync mode and then reading the
recipients gives partial or no results.

The insert will assign the primary key value to the recipient entities.
Therefore we can skip reading the data.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>