Fix collecting addresses in the background
Nextcloud can inject the userId but only for web processes. Now that
messages are sent also in the background and not just from a web request,
the user ID can be null and make the address collection log an
exception.

Following the pattern that user ID should only be injected into a
controller, the user ID is now passed as argument of the collector,
making it suitable for background jobs too.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>