Fix download all as zip
Horde_Nls is listed as optional dependency for horde/mime but unavailable via composer.
Horde_Mime_Part.getSize is uses a class from Horde_Nls without a runtime check if the dependency is available.

Horde_Mime_Part.getSize will return the size of the attachment as KB.
For ZipResponse we need the size in bytes anyway so this patch switch getSize to getBytes to bypass the Horde_Nls dependency.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>