Do not chunk changed/vanished UIDS with QRESYNC
QRESYNC doesn't need the known UIDs as input. In fact it will ignore
the option. As a result the sync will already give an efficient diff of
what really changed. If we still chunk the call it just means that we
will do the exact same operation n times. E.g. on a 600k messages
mailbox the same sync operations is run 60 times.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>