Adjust editor mode on alias change / Add warning on disable formatting
1) Adjust editor mode on alias change

Changing an alias is simple when current and new alias use the same editor mode.
Update alias object and trigger the signature plugin.

When the current alias uses text but the new alias richtext we automatically change the editor mode to html.
TextEditor uses :key="editorMode" to assign the instance to an editorMode. On :key change the old instance is destroyed and a new instance created.
This emit another onEditorReady event and will trigger the signature plugin. To replace an existing signature changeSignature is true and set TRIGGER_CHANGE_ALIAS for the signature command.

2) Add warning on disable formatting

When you compose a richtext message and disable formatting a warning is shown that formatting is lost.
Enable formatting ActionCheckbox is now replaced by buttons with icons.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>