Remove unneccessary account check
As we're already joining on the user accounts table, the account doesn't need to be queried again for a single show.

This is a mini performance optimistation.

THe complete account data could also be returned with the JOIN queries as a second
performance improvement, instead of querying the entity fromt he DB again

The update method does just that at the moment.

Signed-off-by: Anna Larch <anna@nextcloud.com>