Fix lost modifications of preview text
Messages are processed one, on demand. The processing result is stored
to the database. In that process we strip invalid characters from the
preview text and cut it to database column width.
If a message is processed and converted into response JSON, the
stripping and cutting didn't have an effect for the in memory objects.
Only messages loaded from the database cache had a stripped and cut json
representation.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>