fix(XOAUTH2): Defer OAUTH account detection
Defer the detection of OAUTH until auto config is done. Only if the IMAP
and SMTP servers are from Google and we have OAUTH credentials then we
try OAUTH. This makes it possible to use Gmail OAUTh with domains other
than `@gmail.com`.

As a side effect we will no longer require a password unless Google
integration is not set up.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>