fix(imap): Chunk UIDs by string length
IMAP sync requires an IMAP FETCH command that fetches specific messages.
If the full mailbox has to be sync'ed, that can be a long list of UIDs.

The synchronizer had already broken UIDs into chunks by a maximum chunk
size. Yet the chunk size does not correlate with the resulting IMAP
FETCH command length.

An advanced chunk function calculates the UID range string and decreases
the chunk size until the UIDs produce a string that fits the given byte
size.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>