Authorize domain to Mailvelope with a single click
Allow a user with a single click to authorize Mailvelope to expose its
API to the current domain. Previously that took several manual steps in
Mailvelopes config options.

This change includes showing a different text in case Mailvelope is
already installed and the domain authorized. That check notices when
Mailvelope gets activated.

This re-introduces the link(s) to be styled as buttons, as it previously
apparently was intended (the styles were present but not applied due to
a missing class-attribute).

Signed-off-by: Pablo Zmdl <pablo@nextcloud.com>