fix(db): Delete recipients without sub query
MySQL does not use indexes for the "dependent" subquery. It's either a
bug or a missing optimization. EXPLAIN for the two queries is fine in
isolation but when you DELETE with subquery neither the inner nor the
outher query use an index. This causes a full table scan on very large
tables.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>